refactor: @rocket-scripts/webpack-dev-server -> @ssen/webpack-dev-server
