+++ src/@rocket-scripts/web/start.tsx
@@ -2,7 +2,7 @@ import { getBrowserslistQuery } from '@rocket-scripts/browserslist';
 import { webpackConfig as webpackReactConfig } from '@rocket-scripts/react-preset';
 import { getWebpackAlias, icuFormat, rocketTitle } from '@rocket-scripts/utils';
 import { observeAliasChange } from '@rocket-scripts/web/utils/observeAliasChange';
-import { devServerStart, DevServerStartParams } from '@rocket-scripts/webpack-dev-server';
+import { devServerStart, DevServerStartParams } from '@ssen/webpack-dev-server';
 import getPort from 'get-port';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import path from 'path';
